examples/streamlit_app/requirements.txt
streamlit==1.24.0
asyncpg==0.24.0
python-dotenv==0.19.0
pymongo==3.12.3
pydantic==1.10.8
fastapi==0.68.0
uvicorn==0.15.0
langchain==0.0.267
PyYAML==5.3.1
python-jose==3.3.0
cachetools==5.0.0
pillow==9.5.0 